pyproject.toml
[project]
name = "langchain-file-authz"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "click>=8.1.7",
    "faiss-cpu>=1.9.0",
    "google>=3.0.0",
    "langchain-community>=0.3.3",
    "langchain-googledrive>=0.3.0",
    "langchain-openai>=0.2.3",
    "langchain>=0.3.4",
    "pangea-sdk>=5.1.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.13.0",
    "ruff>=0.7.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # ruff defaults
    "E4",
    "E7",
    "E9",
    "F",

    "I",   # isort
    "PTH", # flake8-use-pathlib
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
